requirements.txt
amqp==5.1.1
asgiref==3.4.1
backports.entry-points-selectable==1.1.1
billiard==3.6.4.0
celery==5.2.7
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.3
Django==3.2.13
django-admin-interface==0.18.6
django-admin-rangefilter==0.8.1
django-colorfield==0.6.3
django-constance==2.8.0
django-cors-headers==3.13.0
django-debug-toolbar==3.2.4
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-hosts==4.0
django-import-export==2.6.1
django-mail-queue==3.2.4
django-multiselectfield==0.1.12
django-picklefield==3.0.1
django-rest-framework==0.1.0
django-simple-history==3.0.0
django-sql-explorer==2.4.1
djangorestframework==3.12.4
et-xmlfile==1.1.0
filelock==3.4.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.2
kombu==5.2.4
Markdown==3.3.4
markdown2==2.4.1
MarkupPy==1.14
MarkupSafe==2.0.1
mysqlclient==2.0.3
names==0.3.0
odfpy==1.4.1
openpyxl==3.0.9
packaging==21.0
Pillow==9.0.1
platformdirs==2.4.0
prompt-toolkit==3.0.21
protobuf==3.19.4
pyparsing==3.0.0
pytz==2021.3
PyYAML==6.0
redis==3.5.3
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
tablib==3.0.0
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
virtualenv==20.10.0
wcwidth==0.2.5
xlrd==2.0.1
xlwt==1.3.0
